LPS-67296 Move PostgreSQL's insert generation code
